requirements.txt
openai==1.76.0
datasets==3.5.0
google-genai==1.13.0
tiktoken==0.9.0